+++ src/main/scala/edu/ie3/util/scala/quantities/QuantityUtil.scala
@@ -37,6 +37,9 @@ object QuantityUtil {
       unit: javax.measure.Unit[Q]
   ): ComparableQuantity[Q] = Quantities.getQuantity(0, unit)
 
+  @deprecated(
+    "Use reduceOption { (power1, power2) => power1.add(power2) } instead"
+  )
   def add[Q <: Quantity[Q]](
       quantities: Iterable[Quantity[Q]]
   ): Option[Quantity[Q]] = {