+++ src/test/scala/edu/ie3/simona/model/assets/control/QControlSpec.scala
@@ -6,13 +6,12 @@
 
 package edu.ie3.simona.model.assets.control
 
-import java.util
 import edu.ie3.datamodel.models.input.system.characteristic
 import edu.ie3.datamodel.models.input.system.characteristic.{
+  CharacteristicPoint,
   CosPhiP => CosPhiPInput,
   QV => QVInput
 }
-import edu.ie3.datamodel.models.input.system.characteristic.CharacteristicPoint
 import edu.ie3.simona.exceptions.QControlException
 import edu.ie3.simona.model.participant.control.QControl
 import edu.ie3.simona.model.participant.control.QControl.{
@@ -23,17 +22,20 @@ import edu.ie3.simona.model.participant.control.QControl.{
 import edu.ie3.simona.model.system.Characteristic.XYPair
 import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.util.quantities.PowerSystemUnits._
-import edu.ie3.util.quantities.QuantityUtil
-
-import javax.measure.quantity.Dimensionless
+import edu.ie3.util.scala.quantities.{Megavars, ReactivePower}
 import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor2}
+import squants.Each
 import tech.units.indriya.quantity.Quantities._
 
+import java.util
+import javax.measure.quantity.Dimensionless
 import scala.collection.immutable.TreeSet
 
 class QControlSpec extends UnitSpec with TableDrivenPropertyChecks {
 
-  final val defaultTolerance = 1e-12
+  private val defaultTolerance = 1e-12
+  private implicit val reactivePowerTolerance: ReactivePower = Megavars(1e-12)
+  private implicit val tolerance: squants.Dimensionless = Each(1e-12)
 
   val validCosPhiPInput: characteristic.CosPhiP = new CosPhiPInput(
     "cosPhiP:{(0.0,-1.0),(0.5,-0.8),(1.0,-0.2)}"
@@ -46,8 +48,8 @@ class QControlSpec extends UnitSpec with TableDrivenPropertyChecks {
   def createXYPair(
       d1: Double,
       d2: Double
-  ): XYPair[Dimensionless, Dimensionless] = {
-    XYPair(getQuantity(d1, PU), getQuantity(d2, PU))
+  ): XYPair[squants.Dimensionless, squants.Dimensionless] = {
+    XYPair(Each(d1), Each(d2))
   }
 
   "A valid QControl object" should {
@@ -71,7 +73,7 @@ class QControlSpec extends UnitSpec with TableDrivenPropertyChecks {
         )
       intercept[QControlException](
         QControl(invalidInput)
-      ).getMessage shouldBe "Got an invalid definition of fixed power factor: cosPhiFixed{points=[CharacteristicCoordinate{x=1 p.u., y=2 p.u.}, CharacteristicCoordinate{x=3 p.u., y=4 p.u.}]}. It may only contain one coordinate"
+      ).getMessage shouldBe "Got an invalid definition of fixed power factor: cosPhiFixed{points=[CharacteristicCoordinate{x=1.0 p.u., y=2.0 p.u.}, CharacteristicCoordinate{x=3.0 p.u., y=4.0 p.u.}]}. It may only contain one coordinate"
     }
 
     "parse a valid CosPhiFixed correctly" in {
@@ -115,40 +117,26 @@ class QControlSpec extends UnitSpec with TableDrivenPropertyChecks {
     }
 
     "provide correct values when the requested value is part of the containing xy coordinates" in {
-      val requestedValue = getQuantity(0.5, PU)
-      QuantityUtil.isEquivalentAbs(
-        validCosPhiP.cosPhi(requestedValue),
-        getQuantity(-0.8, PU),
-        defaultTolerance
-      ) shouldBe true
+      val requestedValue = Each(0.5)
+      (validCosPhiP.cosPhi(requestedValue) ~= Each(-0.8)) shouldBe true
     }
 
     "provide an interpolated value when the requested value is not part of the containing xy coordinates" in {
-      val requestedValue = getQuantity(0.75, PU)
-      QuantityUtil.isEquivalentAbs(
-        validCosPhiP.cosPhi(requestedValue),
-        getQuantity(-0.5, PU),
-        defaultTolerance
-      ) shouldBe true
+      val requestedValue = Each(0.75)
+
+      (validCosPhiP.cosPhi(requestedValue) ~= Each(-0.5)) shouldBe true
     }
 
     "provide the last known value when the requested value is outside of the containing xy coordinates" in {
-      QuantityUtil.isEquivalentAbs(
-        validCosPhiP.cosPhi(getQuantity(2.0, PU)),
-        getQuantity(-0.2, PU),
-        defaultTolerance
-      ) shouldBe true
 
-      QuantityUtil.isEquivalentAbs(
-        validCosPhiP.cosPhi(getQuantity(-1.0, PU)),
-        getQuantity(-1.0, PU),
-        defaultTolerance
-      ) shouldBe true
+      (validCosPhiP.cosPhi(Each(2.0)) ~= Each(-0.2)) shouldBe true
+
+      (validCosPhiP.cosPhi(Each(-1.0)) ~= Each(-1.0)) shouldBe true
     }
   }
 
   "A valid Q(V) control" should {
-    val qMax = getQuantity(250d, MEGAVAR)
+    val qMax = Megavars(250.0)
 
     "return correct reactive power for a linear function" in {
       val validQV =
@@ -175,11 +163,7 @@ class QControlSpec extends UnitSpec with TableDrivenPropertyChecks {
       )
 
       forAll(testingPoints) { (v: Double, scaleExpected: Double) =>
-        QuantityUtil.isEquivalentAbs(
-          validQV.q(getQuantity(v, PU), qMax),
-          qMax.multiply(scaleExpected),
-          defaultTolerance
-        ) shouldBe true
+        (validQV.q(Each(v), qMax) ~= qMax * scaleExpected) shouldBe true
       }
     }
 
@@ -217,11 +201,7 @@ class QControlSpec extends UnitSpec with TableDrivenPropertyChecks {
       )
 
       forAll(testingPoints) { (v: Double, scaleExpected: Double) =>
-        QuantityUtil.isEquivalentAbs(
-          validQV.q(getQuantity(v, PU), qMax),
-          qMax.multiply(scaleExpected),
-          defaultTolerance
-        ) shouldBe true
+        (validQV.q(Each(v), qMax) ~= qMax * scaleExpected) shouldBe true
       }
     }
   }