+++ src/test/scala/edu/ie3/simona/service/weather/WeatherSourceSpec.scala
@@ -6,7 +6,16 @@
 
 package edu.ie3.simona.service.weather
 
+import edu.ie3.datamodel.io.factory.timeseries.{
+  CosmoIdCoordinateFactory,
+  IconIdCoordinateFactory,
+  IdCoordinateFactory
+}
 import edu.ie3.datamodel.io.source.IdCoordinateSource
+import edu.ie3.simona.exceptions.{
+  InvalidConfigParameterException,
+  ServiceException
+}
 import edu.ie3.datamodel.models.StandardUnits
 import edu.ie3.datamodel.models.timeseries.individual.{
   IndividualTimeSeries,
@@ -31,6 +40,9 @@ import edu.ie3.simona.service.weather.WeatherSource.{
 import edu.ie3.simona.service.weather.WeatherSourceSpec._
 import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.util.geo.{CoordinateDistance, GeoUtils}
+import edu.ie3.util.quantities.QuantityUtil
+import org.locationtech.jts.geom.{Envelope, Point}
+import tech.units.indriya.ComparableQuantity
 import edu.ie3.util.quantities.interfaces.Irradiance
 import edu.ie3.util.quantities.{PowerSystemUnits, QuantityUtil}
 import org.locationtech.jts.geom.Point
@@ -41,11 +53,13 @@ import tech.units.indriya.unit.Units
 
 import java.time.ZonedDateTime
 import java.util
+import java.util.Optional
+import javax.measure.quantity.Length
 import java.util.{Optional, UUID}
 import javax.measure.quantity.{Speed, Temperature}
 import scala.jdk.CollectionConverters._
 import scala.jdk.OptionConverters._
-import scala.util.{Failure, Success}
+import scala.util.{Failure, Success, Try}
 
 class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
   private val coordinate0 = GeoUtils.buildPoint(51.47, 7.41)
@@ -54,22 +68,20 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
     "issue a ServiceException, if there are not enough coordinates available" in {
       DummyWeatherSource.getNearestCoordinatesWithDistances(
         AgentCoordinates(coordinate0.getY, coordinate0.getX),
-        9,
-        Quantities.getQuantity(28, PowerSystemUnits.KILOMETRE)
+        9
       ) match {
         case Failure(exception: ServiceException) =>
-          exception.getMessage shouldBe "There are not enough coordinates for averaging. Found 8 but need 9."
+          exception.getMessage shouldBe "There are not enough coordinates for averaging. Found 8 within the given distance of 400000 m but need 9. Please make sure that there are enough coordinates within the given distance."
         case _ => fail("You shall not pass!")
       }
     }
     "issue a ServiceException, if there are not enough coordinates in max distance available" in {
       DummyWeatherSource.getNearestCoordinatesWithDistances(
         AgentCoordinates(coordinate0.getY, coordinate0.getX),
-        5,
-        Quantities.getQuantity(5, PowerSystemUnits.KILOMETRE)
+        9
       ) match {
         case Failure(exception: ServiceException) =>
-          exception.getMessage shouldBe "There are not enough coordinates within the max coordinate distance of 5 km. Found 4 but need 5."
+          exception.getMessage shouldBe "There are not enough coordinates for averaging. Found 8 within the given distance of 400000 m but need 9. Please make sure that there are enough coordinates within the given distance."
         case _ => fail("You shall not pass!")
       }
     }
@@ -78,8 +90,7 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
       val agentCoordinates = AgentCoordinates(51.3, 7.3)
       DummyWeatherSource.getNearestCoordinatesWithDistances(
         agentCoordinates,
-        4,
-        Quantities.getQuantity(28, PowerSystemUnits.KILOMETRE)
+        4
       ) match {
         case Failure(exception: ServiceException) =>
           exception.getMessage shouldBe "The queried point shall be surrounded by 4 weather coordinates, which are in each quadrant. This is not the case."
@@ -98,8 +109,7 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
 
       DummyWeatherSource.getNearestCoordinatesWithDistances(
         agentCoordinates,
-        4,
-        Quantities.getQuantity(28, PowerSystemUnits.KILOMETRE)
+        4
       ) match {
         case Success(coordinateDistances) =>
           coordinateDistances.size shouldBe 1
@@ -145,8 +155,7 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
 
       DummyWeatherSource.getNearestCoordinatesWithDistances(
         agentCoordinates,
-        4,
-        Quantities.getQuantity(28, PowerSystemUnits.KILOMETRE)
+        4
       ) match {
         case Success(coordinateDistances) =>
           coordinateDistances.size shouldBe 4
@@ -254,13 +263,12 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
       /* Query more coordinates, than are apparent */
       DummyWeatherSource.getWeightedCoordinates(
         AgentCoordinates(coordinate0.getY, coordinate0.getX),
-        9,
-        Quantities.getQuantity(28, PowerSystemUnits.KILOMETRE)
+        9
       ) match {
         case Failure(exception: ServiceException) =>
           exception.getMessage shouldBe "Determination of coordinate weights failed."
           exception.getCause shouldBe ServiceException(
-            "There are not enough coordinates for averaging. Found 8 but need 9."
+            "There are not enough coordinates for averaging. Found 8 within the given distance of 400000 m but need 9. Please make sure that there are enough coordinates within the given distance."
           )
         case _ => fail("You shall not pass!")
       }
@@ -271,8 +279,7 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
 
       DummyWeatherSource.getWeightedCoordinates(
         agentCoordinates,
-        4,
-        Quantities.getQuantity(28, PowerSystemUnits.KILOMETRE)
+        4
       ) match {
         case Success(WeightedCoordinates(weighting)) =>
           weighting.size shouldBe 1
@@ -300,8 +307,7 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
 
       DummyWeatherSource.getWeightedCoordinates(
         agentCoordinates,
-        4,
-        Quantities.getQuantity(28, PowerSystemUnits.KILOMETRE)
+        4
       ) match {
         case Success(WeightedCoordinates(weighting)) =>
           weighting.corresponds(expectedWeighting) {
@@ -315,6 +321,41 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
             "Determining the nearest weighted coordinates was meant to succeed.",
             exception
           )
+      }
+    }
+
+    "return correct coordinate factory" in {
+      val checkCoordinateFactory =
+        PrivateMethod[IdCoordinateFactory](Symbol("checkCoordinateFactory"))
+
+      val cases = Table(
+        ("gridModel", "expectedClass", "failureMessage"),
+        (
+          "",
+          classOf[InvalidConfigParameterException],
+          "No grid model defined!"
+        ),
+        ("icon", classOf[IconIdCoordinateFactory], ""),
+        ("cosmo", classOf[CosmoIdCoordinateFactory], ""),
+        (
+          "else",
+          classOf[InvalidConfigParameterException],
+          "Grid model 'else' is not supported!"
+        )
+      )
+
+      forAll(cases) { (gridModel, expectedClass, failureMessage) =>
+        val actual =
+          Try(WeatherSource invokePrivate checkCoordinateFactory(gridModel))
+
+        actual match {
+          case Success(factory) =>
+            factory.getClass shouldBe expectedClass
+
+          case Failure(exception) =>
+            exception.getClass shouldBe expectedClass
+            exception.getMessage shouldBe failureMessage
+        }
       }
     }
 
@@ -407,11 +448,11 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
 
       forAll(cases) {
         (
-            timeBasedValues,
-            expectedDiffIrr,
-            expectedDirIrr,
-            expectedTemp,
-            expectedWindVel
+          timeBasedValues,
+          expectedDiffIrr,
+          expectedDirIrr,
+          expectedTemp,
+          expectedWindVel
         ) =>
           val timeSeries: IndividualTimeSeries[WeatherValue] =
             buildTimeSeries(timeBasedValues)
@@ -519,11 +560,11 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
 
       forAll(cases) {
         (
-            timeBasedValues,
-            expectedDiffIrr,
-            expectedDirIrr,
-            expectedTemp,
-            expectedWindVel
+          timeBasedValues,
+          expectedDiffIrr,
+          expectedDirIrr,
+          expectedTemp,
+          expectedWindVel
         ) =>
           val timeSeries: IndividualTimeSeries[WeatherValue] =
             buildTimeSeries(timeBasedValues)
@@ -567,6 +608,7 @@ class WeatherSourceSpec extends UnitSpec with TableDrivenPropertyChecks {
       }
     }
 
+  }
   }
 }
 
@@ -689,6 +731,8 @@ case object WeatherSourceSpec {
   case object DummyWeatherSource extends WeatherSource {
     override protected val idCoordinateSource: IdCoordinateSource =
       DummyIdCoordinateSource
+    override protected val maxCoordinateDistance: ComparableQuantity[Length] =
+      Quantities.getQuantity(400000, Units.METRE)
 
     /** Get the weather data for the given tick as a weighted average taking
       * into account the given weighting of weather coordinates.
@@ -755,5 +799,33 @@ case object WeatherSourceSpec {
 
     override def getAllCoordinates: util.Collection[Point] =
       idToCoordinate.values.toVector.asJava
+
+    def getClosestCoordinates(
+        coordinate: Point,
+        n: Int,
+        distance: ComparableQuantity[Length]
+    ): util.List[CoordinateDistance] = {
+      val points: Set[Point] = coordinateToId.keySet
+
+      val envelope: Envelope =
+        GeoUtils.calculateBoundingBox(coordinate, distance)
+
+      val reducedPoints: Set[Point] = points.flatMap { point =>
+        if (envelope.contains(point.getCoordinate)) {
+          Some(point)
+        } else {
+          None
+        }
+      }
+
+      calculateCoordinateDistances(coordinate, n, reducedPoints.asJava)
+    }
+
+    override def getNearestCoordinates(
+        coordinate: Point,
+        n: Int
+    ): util.List[CoordinateDistance] = {
+      calculateCoordinateDistances(coordinate, n, coordinateToId.keySet.asJava)
+    }
   }
 }